+++ heima-leadnews-service/heima-leadnews-wemedia/src/main/java/com/heima/wemedia/service/WmNewsService.java
@@ -20,7 +20,7 @@ public interface WmNewsService extends IService<WmNews> {
 
     ResponseResult getDetail(Long id);
 
-    ResponseResult deleteNews(Long id);
+    ResponseResult deleteNews(Integer id);
 
     ResponseResult downOrUpNews(WmNewsDto wmNewsDto);
 }
